requirements.txt
# python requirements
# Python version 3.7.6
matplotlib
numpy
pandas
pytorch
scikit-learn
scipy

